\name{buildArchive}
\alias{buildArchive}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Build a TCGA archive with aux, Level 1, Level 2, Level 3 and mage-tab directories
}
\description{
The function will create the necessary directories and write out the appropriate files which are required as part of a TCGA archive. Note that the current implementation is tailored to work with a directory structure that has been created on USC's High Performance Computing Cluster \acronym{HPCC}.
}
\usage{
buildArchive(map, old.version = "0", new.version = "0", base = NULL, platform = "HumanMethylation450", magetab.version = NULL, write.magetab = TRUE, lvls = c(1:3), revision = FALSE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{map}{
	A MethyLumiSet object or a data frame containing metadata about the TCGA archive being built
}
  \item{old.version}{
	The current version of the archive on the TCGA data portal
}
  \item{new.version}{
	The new version of the archive that will be created for upload to the TCGA data portal
}
  \item{base}{
	The path to the base directory where the archive will be created
}
  \item{platform}{
	The assay platform for which the archive is being created
}
  \item{magetab.version}{
	The version of the mage-tab directory to create. If \code{NULL} defaults to \code{new.version}
}
  \item{write.magetab}{
	Logical indicating whether or not the mage-tab directory should be created
}
  \item{lvls}{
	Integer vector specifying the Levels for which to create directories
}
  \item{revision}{
	Logical indicating whether this is build is a revision or a fresh build
}
}
\details{
The function assumes that there are symlinks \emph{$HOME/meth27k} and \emph{$HOME/meth450k} that point to the location of the \bold{27k} and \bold{450k} platform specific archive directories. This assumption is taken into account when the \code{base} option is \code{NULL}.

The minimal directory structure for both the \bold{27k} and \bold{450k} platforms consists of a parent \emph{meth27k} or \emph{meth450k} directory which in turn each contain the following directories:

\itemize{
  \item \emph{raw} contains the raw \bold{idat} files generated by the \bold{iScan} system and are organized by Beadchip. This directory also contains disease specific directories which are generated by the \code{linkRawData} function

  \item \emph{tcga} contains the packaged directories each set of which is under the version for which they were generated

  \item \emph{MethyLumiSets} contains raw and normalized MethyLumiSets for each tumor type
  }
}

\author{
Tim Triche Jr. and Moiz Bootwalla
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
\dontrun{
	buildArchive(map, old.version='3', new.version='4')
}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
