\name{mapBatches}
\alias{mapBatches}
\title{
  Given a pile of Dan's Excel files, attempt to merge them into a map.
}
\description{
  If the function squawks, do what it says, and fix the offending elements.
}
\author{
Tim Triche Jr.
}
\examples{
\dontrun{
     map <- mapBatches(list.files(patt='xls'))
}
}

